# Adds ground work for custom blueprints. Fixes #660