--- a/constraint/bls12-377/solver.go
+++ b/constraint/bls12-377/solver.go
@@ -358,6 +358,9 @@ func (s *solver) IsSolved(vID uint32) bool {
 // evaluates it and return the result and the number of uint32 word read.
 func (s *solver) Read(calldata []uint32) (constraint.Element, int) {
 	if s.Type == constraint.SystemSparseR1CS {
+		if calldata[0] != 1 {
+			panic("invalid calldata")
+		}
 		return s.GetValue(calldata[1], calldata[2]), 3
 	}
 	var r fr.Element