--- a/constraint/bls12-381/solver.go
+++ b/constraint/bls12-381/solver.go
@@ -376,10 +376,10 @@ func (s *solver) Read(calldata []uint32) (constraint.Element, int) {
 
 // processInstruction decodes the instruction and execute blueprint-defined logic.
 // an instruction can encode a hint, a custom constraint or a generic constraint.
-func (solver *solver) processInstruction(inst constraint.PackedInstruction, scratch *scratch) error {
+func (solver *solver) processInstruction(pi constraint.PackedInstruction, scratch *scratch) error {
 	// fetch the blueprint
-	blueprint := solver.Blueprints[inst.BlueprintID]
-	calldata := solver.GetCallData(inst)
+	blueprint := solver.Blueprints[pi.BlueprintID]
+	inst := pi.Unpack(&solver.System)
 	cID := inst.ConstraintOffset // here we have 1 constraint in the instruction only
 
 	if solver.Type == constraint.SystemR1CS {
@@ -387,13 +387,13 @@ func (solver *solver) processInstruction(inst constraint.PackedInstruction, scra
 			// TODO @gbotrel we use the solveR1C method for now, having user-defined
 			// blueprint for R1CS would require constraint.Solver interface to add methods
 			// to set a,b,c since it's more efficient to compute these while we solve.
-			bc.DecompressR1C(&scratch.tR1C, calldata)
+			bc.DecompressR1C(&scratch.tR1C, inst)
 			return solver.solveR1C(cID, &scratch.tR1C)
 		}
 	} else if solver.Type == constraint.SystemSparseR1CS {
 		// blueprint declared "I know how to solve this."
 		if bc, ok := blueprint.(constraint.BlueprintSolvable); ok {
-			if err := bc.Solve(solver, calldata); err != nil {
+			if err := bc.Solve(solver, inst); err != nil {
 				return solver.wrapErrWithDebugInfo(cID, err)
 			}
 			return nil
@@ -403,7 +403,7 @@ func (solver *solver) processInstruction(inst constraint.PackedInstruction, scra
 	// blueprint encodes a hint, we execute.
 	// TODO @gbotrel may be worth it to move hint logic in blueprint "solve"
 	if bc, ok := blueprint.(constraint.BlueprintHint); ok {
-		bc.DecompressHint(&scratch.tHint, calldata)
+		bc.DecompressHint(&scratch.tHint, inst)
 		return solver.solveWithHint(&scratch.tHint)
 	}